backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
google-generativeai
python-dotenv